requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
idna==3.3
numpy==1.22.3
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
